add helper for generating block file
